Trying to finish 2b
I've tried to solve for the right potential and matrix. I've made the algoritm to find how many mesh points that are nessasary, but the eigenvalues coming out are not correct. I want to check if the function still solves correctly so I am branching out to be able to change the structure without loosing anything.